Module jpeg_decoder::decoder

Modules

lossless 🔒

Structs

Decoder

JPEG decoder

ImageInfo

Represents metadata of an image.

Enums

ColorTransform

Describes the colour transform to apply before binary data is returned

PixelFormat

An enumeration over combinations of color spaces and bit depths a pixel can have.

Constants

FIXED_POINT_OFFSET 🔒

HALF 🔒

MAX_COMPONENTS

Statics

UNZIGZAG 🔒

Functions

choose_color_convert_func 🔒

clamp_fixed_point 🔒

color_convert_line_cmyk 🔒

color_convert_line_rgb 🔒

color_convert_line_ycbcr 🔒

color_convert_line_ycck 🔒

color_no_convert 🔒

compute_image 🔒

decode_block 🔒

decode_block_successive_approximation 🔒

refine_non_zeroes 🔒

stbi_f2f 🔒

ycbcr_to_rgb 🔒