Enum jpeg_decoder::UnsupportedFeature

pub enum UnsupportedFeature {
    Hierarchical,
    Lossless,
    ArithmeticEntropyCoding,
    SamplePrecision(u8),
    ComponentCount(u8),
    DNL,
    SubsamplingRatio,
    NonIntegerSubsamplingRatio,
}

An enumeration over JPEG features (currently) unsupported by this library.

Support for features listed here may be included in future versions of this library.

Variants
Hierarchical

Hierarchical JPEG.

Lossless

Lossless JPEG.

ArithmeticEntropyCoding

JPEG using arithmetic entropy coding instead of Huffman coding.

SamplePrecision(u8)

Sample precision in bits. 8 bit sample precision is what is currently supported.

ComponentCount(u8)

Number of components in an image. 1, 3 and 4 components are currently supported.

DNL

An image can specify a zero height in the frame header and use the DNL (Define Number of Lines) marker at the end of the first scan to define the number of lines in the frame.

SubsamplingRatio

Subsampling ratio.

NonIntegerSubsamplingRatio

A subsampling ratio not representable as an integer.