Module jpeg_decoder::worker

Re-exports

pub use self::multithreaded::MultiThreadedWorker as PlatformWorker;

Modules

immediate

multithreaded

This module implements per-component parallelism. It should be possible to implement per-row parallelism as well, which should also boost performance of grayscale images and allow scaling to more cores. However, that would be more complex, so we use this as a starting point.

Structs

RowData

Traits

Worker