Crate jpeg_decoder

This crate contains a JPEG decoder.

Examples

use jpeg_decoder::Decoder;
use std::fs::File;
use std::io::BufReader;

let file = File::open("tests/reftest/images/extraneous-data.jpg").expect("failed to open file");
let mut decoder = Decoder::new(BufReader::new(file));
let pixels = decoder.decode().expect("failed to decode image");
let metadata = decoder.info().unwrap();

Get metadata from a file without decoding it:

use jpeg_decoder::Decoder;
use std::fs::File;
use std::io::BufReader;

let file = File::open("tests/reftest/images/extraneous-data.jpg").expect("failed to open file");
let mut decoder = Decoder::new(BufReader::new(file));
decoder.read_info().expect("failed to read metadata");
let metadata = decoder.info().unwrap();

Structs

Decoder

JPEG decoder

ImageInfo

Represents metadata of an image.

Enums

CodingProcess

Represents the coding process of an image.

ColorTransform

Describes the colour transform to apply before binary data is returned

Error

Errors that can occur while decoding a JPEG image.

PixelFormat

An enumeration over combinations of color spaces and bit depths a pixel can have.

UnsupportedFeature

An enumeration over JPEG features (currently) unsupported by this library.