Module jpeg_decoder::decoder

Modules

lossless ğŸ”’

Structs

Decoder

JPEG decoder

ImageInfo

Represents metadata of an image.

Enums

ColorTransform

Describes the colour transform to apply before binary data is returned

PixelFormat

An enumeration over combinations of color spaces and bit depths a pixel can have.

Constants

FIXED_POINT_OFFSET ğŸ”’

HALF ğŸ”’

MAX_COMPONENTS

Statics

UNZIGZAG ğŸ”’

Functions

choose_color_convert_func ğŸ”’

clamp_fixed_point ğŸ”’

color_convert_line_cmyk ğŸ”’

color_convert_line_rgb ğŸ”’

color_convert_line_ycbcr ğŸ”’

color_convert_line_ycck ğŸ”’

color_no_convert ğŸ”’

compute_image ğŸ”’

decode_block ğŸ”’

decode_block_successive_approximation ğŸ”’

refine_non_zeroes ğŸ”’

stbi_f2f ğŸ”’

ycbcr_to_rgb ğŸ”’