Struct jpeg_decoder::Decoder

pub struct Decoder<R> { /* private fields */ }

JPEG decoder

Implementations

impl<R: Read> Decoder<R>

pub fn new(reader: R) -> Decoder<R>

Creates a new Decoder using the reader reader.

pub fn info(&self) -> Option<ImageInfo>

Returns metadata about the image.

The returned value will be None until a call to either read_info or decode has returned Ok.

pub fn icc_profile(&self) -> Option<Vec<u8>>

Returns the embeded icc profile if the image contains one.

pub fn read_info(&mut self) -> Result<(), Error>

Tries to read metadata from the image without decoding it.

If successful, the metadata can be obtained using the info method.

pub fn scale(
&mut self,
requested_width: u16,
requested_height: u16
) -> Result<(u16, u16), Error>

Configure the decoder to scale the image during decoding.

This efficiently scales the image by the smallest supported scale factor that produces an image larger than or equal to the requested size in at least one axis. The currently implemented scale factors are 1/8, 1/4, 1/2 and 1.

To generate a thumbnail of an exact size, pass the desired size and then scale to the final size using a traditional resampling algorithm.

pub fn decode(&mut self) -> Result<Vec<u8>, Error>

Decodes the image and returns the decoded pixels if successful.